+++ src/main/java/eu/manuelgu/discordmc/MessageAPI.java
@@ -1,19 +1,18 @@
 package eu.manuelgu.discordmc;
 
 import com.vdurmont.emoji.EmojiParser;
-
 import org.bukkit.Bukkit;
 import org.bukkit.ChatColor;
-
 import sx.blah.discord.handle.obj.IChannel;
 import sx.blah.discord.util.DiscordException;
-import sx.blah.discord.util.HTTP429Exception;
 import sx.blah.discord.util.MessageBuilder;
 import sx.blah.discord.util.MissingPermissionsException;
+import sx.blah.discord.util.RequestBuffer;
+
+import java.util.List;
 
 public class MessageAPI {
     private static DiscordMC plugin;
-    private static IChannel channel;
 
     public MessageAPI(DiscordMC main) {
         MessageAPI.plugin = main;
@@ -22,15 +21,18 @@ public class MessageAPI {
     /**
      * Send a chat message to the Minecraft server
      *
+     * @param origin   {@link IChannel} the message came from
      * @param username player who sent the message
      * @param message  the actual message which gets sent
      */
-    public static void sendToMinecraft(String username, String message) {
-        String formattedMessage = DiscordMC.get().getConfig().getString("settings.templates.chat_message_minecraft")
-                .replaceAll("%u", username)
-                .replaceAll("%m", message);
+    public static void sendToMinecraft(IChannel origin, String username, String message) {
+        String formattedMessage = ChatColor.translateAlternateColorCodes('&', DiscordMC.get().getConfig().getString("settings.templates.chat_message_minecraft")
+                .replace("%user", username)
+                .replace("%channel", origin.getName()));
 
-        Bukkit.broadcastMessage(ChatColor.translateAlternateColorCodes('&', EmojiParser.parseToAliases(formattedMessage)));
+        DiscordMC.getSubscribedPlayers().forEach(uuid -> Bukkit.getPlayer(uuid).sendMessage(
+                EmojiParser.parseToAliases(formattedMessage
+                        .replaceAll("%message", ChatColor.stripColor(message)))));
     }
 
     /**
@@ -51,23 +53,35 @@ public class MessageAPI {
         if (!DiscordMC.getClient().isReady()) {
             return;
         }
-        Bukkit.getServer().getScheduler().runTaskAsynchronously(plugin, () -> {
-            channel = DiscordMC.getChannel();
+        Bukkit.getServer().getScheduler().runTaskAsynchronously(plugin, () -> DiscordMC.getMinecraftToDiscord().forEach(channel -> sendToDiscord(channel, message)));
+    }
 
-            if (channel == null) {
-                plugin.getLogger().severe("Could not send message, channel was null");
-                return;
-            }
-
+    /**
+     * Send a message to a specific channel
+     *
+     * @param channel channel to receive message
+     * @param message message to send
+     */
+    public static void sendToDiscord(IChannel channel, String message) {
+        RequestBuffer.request(() -> {
             try {
                 new MessageBuilder(DiscordMC.getClient()).appendContent(message).withChannel(channel).build();
-            } catch (DiscordException e) {
-                plugin.getLogger().severe("Discord threw an exception while sending the message. " + e.getErrorMessage());
-            } catch (HTTP429Exception ignored) {
+            } catch (DiscordException ignored) {
             } catch (MissingPermissionsException e) {
-                plugin.getLogger().severe("Your Bot is missing required permission to perform this action! "
+                plugin.getLogger().severe("Your Bot is missing required permissions to perform this action! "
                         + e.getErrorMessage());
             }
+            return null;
         });
+    }
+
+    /**
+     * Send a message to a list of specific channels
+     *
+     * @param channels channels that receive message
+     * @param message  message to send
+     */
+    public static void sendToDiscord(List<IChannel> channels, String message) {
+        channels.forEach(channel -> sendToDiscord(channel, message));
     }
 }